Port Hub V3 (imxdimagic) platform to Yocto Scarthgap with hardware fixes
Bring up the i.MX6DL-based Hub V3 on Scarthgap (kernel 6.6) with fixes
for networking, sound, and serial ports discovered during bringup:

- Add kernel 6.6 DTS, defconfig, and custom driver patches for Hub V3
- Fix RGMII: add qca,clk-out-frequency to AR8035 PHY node so at803x
  driver re-enables 125MHz clock after kernel PHY reset
- Fix sound: use NAU8810 PLL mode instead of MCLK prescaler — 24MHz
  CKO1 cannot produce exact audio clocks via prescaler alone
- Fix sound: restore missing TDM slot configuration on CPU DAI
- Fix serial: move ttyO* symlinks into platform-specific irisinit
  scripts so irisagent no longer overwrites i.MX ttymxc symlinks
- Fix DHCP: add ifplugd race workaround to irisinit-imxdimagic
- Update machine config, build tools, submodules, and recipe fixes
  for Scarthgap compatibility (license syntax, compiler warnings,
  bluez5 GATT patch no longer needed)

+++ meta-iris/recipes-core/iris-utils/files/irisinitd.c
@@ -396,8 +396,8 @@ static gboolean pwrDownButtonHandler(GIOChannel *channel, GIOCondition cond,
         time_t delta = time(NULL) - pwrDownTime;
         if (delta >= POWER_DOWN_PERIOD) {
             syslog(LOG_ERR,
-                   "Power down button pushed for %ld seconds - halting!",
-                   delta);
+                   "Power down button pushed for %lld seconds - halting!",
+                   (long long)delta);
 
             // Set LED pattern then wait (5 seconds in total)
             setLedMode("turning-off");
